[gui][stats] merge in the same file StatsLineWidget and StatsTable
